Fix a multi-threading deadlock.
As we're not going to the lengths of using separate input and output
threads, and flip-flopping between dispatching jobs and querying the
results, we can get to a block on the dispatcher as the output queue
is full, but being blocked we then never drain it.

The dispatch now never blocks (going over-sized), but the result
draining may be blocking if the input queue is full.

Also add locking to codec metrics.